requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
python-dotenv==1.0.1
lxml==5.1.0
